+++ dashboard/src/pages/dashboard/index.tsx
@@ -11,6 +11,8 @@ import DashboardLayout from '@/layouts/DashboardLayout';
 import { useGetDashboardData } from '@/services/api';
 import { getQueryData } from '@/utils/query';
 import { useProcessData } from '@/hooks/useProcessData';
+import { HalfContainer, HalfContainerCard } from '@/components/containers';
+import { Container, Text } from '@nextui-org/react';
 
 const Dashboard: NextPageWithLayout = () => {
     const { data, isLoading } = useGetDashboardData();
@@ -19,7 +21,7 @@ const Dashboard: NextPageWithLayout = () => {
 
     console.log('dashboardData:', data);
 
-    const { commandsRan, users, servers, songsPlayed } =
+    const { commandsRan, users, servers, songsPlayed, nodes } =
         getQueryData(data) || {};
 
     return (
@@ -29,33 +31,97 @@ const Dashboard: NextPageWithLayout = () => {
             </Head>
 
             <h1>Dashboard</h1>
-            <div
+            <Container
                 style={{
                     display: 'flex',
+                    flexDirection: 'column',
+                    gap: '20px',
                 }}
             >
-                <StatCard
-                    title="Commands Ran"
-                    amount={processData(commandsRan)}
-                    icon={<RocketLaunchRounded fontSize="large" />}
-                />
-                <StatCard
-                    title="Users"
-                    amount={processData(users)}
-                    icon={<PersonRounded fontSize="large" />}
-                />
-                <StatCard
-                    title="Servers"
-                    amount={processData(servers)}
-                    icon={<DnsRounded fontSize="large" />}
-                />
+                <HalfContainer>
+                    <StatCard
+                        title="Commands Ran"
+                        amount={processData(commandsRan)}
+                        icon={<RocketLaunchRounded fontSize="large" />}
+                    />
+                    <StatCard
+                        title="Users"
+                        amount={processData(users)}
+                        icon={<PersonRounded fontSize="large" />}
+                    />
+                    <StatCard
+                        title="Servers"
+                        amount={processData(servers)}
+                        icon={<DnsRounded fontSize="large" />}
+                    />
 
-                <StatCard
-                    title="Songs Played"
-                    amount={processData(songsPlayed)}
-                    icon={<AudiotrackRounded fontSize="large" />}
-                />
-            </div>
+                    <StatCard
+                        title="Songs Played"
+                        amount={processData(songsPlayed)}
+                        icon={<AudiotrackRounded fontSize="large" />}
+                    />
+
+                </HalfContainer>
+                <HalfContainer>
+                    <HalfContainerCard>
+                    <Text h4 css={ {color: 'GrayText'} }>Nodes</Text>
+                        <div>
+                            <div
+                                style={{
+                                    display: 'flex',
+                                    flexDirection: 'row',
+                                    flexWrap: 'wrap',
+                                    gap: '10px',
+                                }}
+                            >
+                                {
+                                    nodes &&
+                                    Object.entries(nodes).map(
+                                        ([nodeId, node]) => (
+                                            <div key={nodeId}
+                                                style={{
+                                                    display: 'flex',
+                                                    flexDirection: 'column',
+                                                    justifyContent: 'center',
+                                                }}
+                                            >
+                                                <span style={{ fontSize: '16px', fontWeight: 600 }}>{nodeId}</span>
+                                                <div
+                                                    style={{
+                                                        backgroundColor: '#121212',
+                                                        borderRadius: '5px',
+                                                        width: '100%',
+                                                    }}
+                                                >
+                                                    <pre
+                                                        style={{
+                                                            fontSize: '12px',
+                                                            color: node.connected ? 'green' : 'red',
+                                                        }}
+                                                    >
+                                                        <div>
+                                                            Uptime: {node.stats.uptime}
+                                                        </div>
+                                                        <div>
+                                                            RAM: {node.stats.ramUsage} MB / {node.stats.ramTotal} MB
+                                                        </div>
+                                                        <div>
+                                                            CPU: {node.stats.cores === 1 ? '1 Core' : `${node.stats.cores} Cores`}
+                                                        </div>
+                                                        <div>
+                                                            Players: {node.nodeStats.playingPlayers}/{node.nodeStats.players}
+                                                        </div>
+                                                    </pre>
+                                                </div>
+                                            </div>
+                                        )
+                                    )
+                                }
+                            </div>
+                        </div>
+                    </HalfContainerCard>
+                </HalfContainer>
+            </Container>
         </>
     );
 };
