+++ djs-bot/api/v1/src/routes/v1/routesHandler/servers.ts
@@ -4,8 +4,6 @@ import { RouteHandler } from '../../../interfaces/common';
 import { createReply } from '../../../utils/reply';
 import { getUserGuilds } from '../../../services/discord';
 import { Guild } from 'discord.js';
-import moment from 'moment';
-
 
 const handler: RouteHandler = async (request, reply) => {
   const bot = getBot();
@@ -16,34 +14,6 @@ const handler: RouteHandler = async (request, reply) => {
       APIError.STATUS_CODES.NOT_FOUND,
       APIError.ERROR_CODES.NOT_FOUND,
     );
-
-  const lavaNodes = bot.manager?.Engine.nodes.entries();
-
-  if (!lavaNodes)
-    throw new APIError(
-      'Lava nodes not found',
-      APIError.STATUS_CODES.NOT_FOUND,
-      APIError.ERROR_CODES.NOT_FOUND,
-    );
-
-  let lavaNodesAggregates = {};
-  for (const [nodeId, node] of lavaNodes) {
-    lavaNodesAggregates = {
-      ...lavaNodesAggregates,
-      [nodeId]: {
-        id: nodeId,
-        connected: node.connected,
-        nodeStats: node.stats,
-        stats: {
-          cores: node.stats.cpu.cores,
-          // @ts-ignore
-          uptime: moment.duration(node.stats.uptime).format("d[ Days]・h[ Hrs]・m[ Mins]・s[ Secs]"),
-          ramUsage: (node.stats.memory.used / 1024 / 1024).toFixed(2),
-          ramTotal: (node.stats.memory.allocated / 1024 / 1024).toFixed(2),
-        }
-      },
-    };
-  }
 
   const id = (request.query as { id: string | undefined }).id;
   if (id) {
@@ -96,7 +66,6 @@ const handler: RouteHandler = async (request, reply) => {
             ?.duration,
         },
       },
-      nodes: lavaNodesAggregates,
     });
   }
